\name{esmpack-package}
\alias{esmpack-package}
\alias{esmpack}
\docType{package}
\title{esmpack: Functions that Facilitate Preparation and Management of ESM/EMA Data}
\description{
   The \pkg{esmpack} package is a collection of functions that facilitate preparation, management, visualization, and analysis of data collected via the experience sampling method (ESM) and ecological momentary assessment (EMA).
}
\author{
   Wolfgang Viechtbauer \email{wvb@wvbauer.com} \cr
   author homepage: \url{https://www.wvbauer.com/} \cr
}
\keyword{package}
