\name{get.day.number}
\alias{get.day.number}
\title{Construct Day Numbers from Notification Numbers}
\description{Function to construct the day numbers based on the notification numbers.}
\usage{
get.day.number(notification.number, beeps.per.day = 10, total.notifications = 50)
}
\arguments{
   \item{notification.number}{blah}
   \item{beeps.per.day}{blah}
   \item{total.notifications}{blah}
}
\details{
   The function can be used to ...
}
\value{
   The function returns a vector.
}
\author{
   Mihai Constantin \email{m.a.constantin@uvt.nl} \cr
}
\examples{
# add an example
}
\keyword{manip}
